# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Xeno Sir AI Tutor
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserInfoRole = (typeof UserInfoRole)[keyof typeof UserInfoRole];

export const UserInfoRole = {
  admin: "admin",
  student: "student",
} as const;

export interface UserInfo {
  id: number;
  username: string;
  role: UserInfoRole;
  displayName: string;
}

export interface LoginResponse {
  user: UserInfo;
  token: string;
}

export interface Lecture {
  id: number;
  lectureNumber: number;
  title: string;
  transcript: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateLectureBody {
  lectureNumber: number;
  title: string;
  transcript: string;
}

export interface UpdateLectureBody {
  lectureNumber?: number;
  title?: string;
  transcript?: string;
}

export interface Student {
  id: number;
  username: string;
  displayName: string;
  createdAt: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface StudentWithConversations {
  id: number;
  username: string;
  displayName: string;
  createdAt: string;
  conversations: OpenaiConversation[];
}

export interface CreateStudentBody {
  username: string;
  password: string;
  displayName: string;
}

/**
 * Response language preference
 */
export type AskXenoSirBodyLanguage =
  (typeof AskXenoSirBodyLanguage)[keyof typeof AskXenoSirBodyLanguage];

export const AskXenoSirBodyLanguage = {
  hindi: "hindi",
  english: "english",
  auto: "auto",
} as const;

export interface AskXenoSirBody {
  question: string;
  /** Optional - specific lecture IDs to focus on. If empty, uses all lectures. */
  lectureIds?: number[];
  /** Optional - continue an existing conversation */
  conversationId?: number;
  /** Response language preference */
  language?: AskXenoSirBodyLanguage;
}

export type GenerateExamBodyDifficulty =
  (typeof GenerateExamBodyDifficulty)[keyof typeof GenerateExamBodyDifficulty];

export const GenerateExamBodyDifficulty = {
  basic: "basic",
  medium: "medium",
  advanced: "advanced",
  extreme: "extreme",
} as const;

export type GenerateExamBodyExamType =
  (typeof GenerateExamBodyExamType)[keyof typeof GenerateExamBodyExamType];

export const GenerateExamBodyExamType = {
  mcq: "mcq",
  subjective: "subjective",
  mixed: "mixed",
} as const;

export interface GenerateExamBody {
  /** Lecture IDs to generate exam from. If empty, uses all lectures. */
  lectureIds?: number[];
  difficulty: GenerateExamBodyDifficulty;
  examType: GenerateExamBodyExamType;
  /**
   * @minimum 5
   * @maximum 50
   */
  questionCount: number;
}

export type ExamQuestionType =
  (typeof ExamQuestionType)[keyof typeof ExamQuestionType];

export const ExamQuestionType = {
  mcq: "mcq",
  subjective: "subjective",
} as const;

export interface ExamQuestion {
  questionNumber: number;
  question: string;
  type: ExamQuestionType;
  /** For MCQ only */
  options?: string[];
  correctAnswer: string;
  explanation: string;
}

export interface ExamResponse {
  lecturesCovered: string[];
  difficulty: string;
  questions: ExamQuestion[];
  totalQuestions: number;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}