# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Xeno Sir AI Tutor
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login as admin or student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /lectures:
    get:
      operationId: listLectures
      tags: [lectures]
      summary: List all lectures
      responses:
        "200":
          description: List of lectures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lecture"
    post:
      operationId: createLecture
      tags: [lectures]
      summary: Create/add a new lecture transcript (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLectureBody"
      responses:
        "201":
          description: Created lecture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lectures/{id}:
    get:
      operationId: getLecture
      tags: [lectures]
      summary: Get a single lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lecture details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateLecture
      tags: [lectures]
      summary: Update a lecture (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLectureBody"
      responses:
        "200":
          description: Updated lecture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLecture
      tags: [lectures]
      summary: Delete a lecture (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /chat/ask:
    post:
      operationId: askXenoSir
      tags: [openai]
      summary: Ask Xeno Sir a question with lecture context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskXenoSirBody"
      responses:
        "200":
          description: SSE stream of response
          content:
            text/event-stream: {}
  /chat/exam:
    post:
      operationId: generateExam
      tags: [openai]
      summary: Generate an exam/quiz from lecture transcripts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateExamBody"
      responses:
        "200":
          description: Generated exam questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResponse"
  /chat/history/{studentId}:
    get:
      operationId: getChatHistory
      tags: [openai]
      summary: Get chat history for a student (admin only)
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserInfo"
        token:
          type: string
      required:
        - user
        - token
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, student]
        displayName:
          type: string
      required:
        - id
        - username
        - role
        - displayName
    Lecture:
      type: object
      properties:
        id:
          type: integer
        lectureNumber:
          type: integer
        title:
          type: string
        transcript:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - lectureNumber
        - title
        - transcript
        - createdAt
        - updatedAt
    CreateLectureBody:
      type: object
      properties:
        lectureNumber:
          type: integer
        title:
          type: string
        transcript:
          type: string
      required:
        - lectureNumber
        - title
        - transcript
    UpdateLectureBody:
      type: object
      properties:
        lectureNumber:
          type: integer
        title:
          type: string
        transcript:
          type: string
...
    AskXenoSirBody:
      type: object
      properties:
        question:
          type: string
        lectureIds:
          type: array
          items:
            type: integer
          description: Optional - specific lecture IDs to focus on. If empty, uses all lectures.
        conversationId:
          type: integer
          description: Optional - continue an existing conversation
        language:
          type: string
          enum: [hindi, english, auto]
          description: Response language preference
      required:
        - question
    GenerateExamBody:
      type: object
      properties:
        lectureIds:
          type: array
          items:
            type: integer
          description: Lecture IDs to generate exam from. If empty, uses all lectures.
        difficulty:
          type: string
          enum: [basic, medium, advanced, extreme]
        examType:
          type: string
          enum: [mcq, subjective, mixed]
        questionCount:
          type: integer
          minimum: 5
          maximum: 50
      required:
        - difficulty
        - examType
        - questionCount
    ExamQuestion:
      type: object
      properties:
        questionNumber:
          type: integer
        question:
          type: string
        type:
          type: string
          enum: [mcq, subjective]
        options:
          type: array
          items:
            type: string
          description: For MCQ only
        correctAnswer:
          type: string
        explanation:
          type: string
      required:
        - questionNumber
        - question
        - type
        - correctAnswer
        - explanation
    ExamResponse:
      type: object
      properties:
        lecturesCovered:
          type: array
          items:
            type: string
        difficulty:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/ExamQuestion"
        totalQuestions:
          type: integer
      required:
        - lecturesCovered
        - difficulty
        - questions
        - totalQuestions
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
...