# lib/api-zod/src/generated/types/examResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Xeno Sir AI Tutor
 * OpenAPI spec version: 0.1.0
 */
import type { ExamQuestion } from "./examQuestion";

export interface ExamResponse {
  lecturesCovered: string[];
  difficulty: string;
  questions: ExamQuestion[];
  totalQuestions: number;
}